requirements.txt
streamlit==1.27.0
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
surprise==0.1
Pillow==9.5.0
mysql-connector-python==8.1.0
mlxtend==0.22.0
matplotlib==3.7.2
seaborn==0.12.2
scikit-surprise
